requirements.txt
requests>=2.31
beautifulsoup4>=4.12
SQLAlchemy>=2.0
typer>=0.12
python-dateutil>=2.9
reportlab>=4.0
pytest>=8.0
feedparser>=6.0
